[Ref] libmodplug emulation: Silence warnings on Win64.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@5154 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ libopenmpt/libopenmpt_modplug_cpp.cpp
@@ -301,7 +301,7 @@ UINT CSoundFile::GetSongComments( LPSTR s, UINT cbsize, UINT linesize ) {
 	}
 	std::strncpy( s, mod->get_metadata("message").c_str(), cbsize );
 	s[ cbsize - 1 ] = '\0';
-	return std::strlen( s ) + 1;
+	return static_cast<UINT>( std::strlen( s ) + 1 );
 }
 
 void CSoundFile::SetCurrentPos( UINT nPos ) {
@@ -400,7 +400,7 @@ UINT CSoundFile::Read( LPVOID lpBuffer, UINT cbBuffer ) {
 			dst[sample] = tmpbuf[sample] << (32-16-1-MIXING_ATTENUATION);
 		}
 	}
-	return frames_rendered * get_frame_size();
+	return static_cast<UINT>( frames_rendered * get_frame_size() );
 }
 
 /*
